fix(tarko): create shared CodeEditorHeader component for consistent file display
